+++ src/index.ts
@@ -1,11 +1,14 @@
 #!/usr/bin/env node
 
 import axios from 'axios';
+import Debug from 'debug';
 import * as fs from 'fs';
 import * as Mustache from 'mustache';
 import * as path from 'path';
 
 (async function main() {
+
+    const debug = Debug('create-package');
 
     interface ITemplateVariables {
         [index: string]: string;
@@ -14,12 +17,17 @@ import * as path from 'path';
     let packageScope = '';
     let packageName = process.argv[2];
 
+    debug(`Got package name argument: ${packageName}`);
+
     const scopeMatchRegex = new RegExp(/@(.+?)\//);
     const privatePackageMatch = packageName.match(scopeMatchRegex);
     if (privatePackageMatch) {
         packageScope = privatePackageMatch[0];
         packageName = packageName.replace(scopeMatchRegex, '');
     }
+
+    debug(`Package scope: ${packageScope}`);
+    debug(`Package name: ${packageName}`);
 
     if (!packageName) {
         console.error('A package name is required!');
@@ -31,12 +39,16 @@ import * as path from 'path';
         process.exit(1);
     }
 
+    debug(`Attempting to create folder: ${packageName}`);
+
     if (fs.existsSync(packageName)) {
         console.error('Directory already exists, cannot create package!');
         process.exit(1);
     }
 
     fs.mkdirSync(packageName);
+
+    debug(`Folder "${packageName}" created.`);
 
     const latestTypesNodeVersion = await getLatestPackageVersion('@types/node');
     const latestTsLintVersion = await getLatestPackageVersion('tslint');
@@ -56,30 +68,51 @@ import * as path from 'path';
     const templateDirectory = path.join(__dirname, '..', 'templates');
 
     async function getLatestPackageVersion(name: string): Promise<string> {
+        debug(`Getting latest version number for package: ${name}`);
+
         const versionToGet = name.startsWith('@') ? '*' : 'latest';
-        const response = await axios.get(`https://registry.npmjs.org/${ encodeURIComponent(name) }/${versionToGet}`);
+
+        const url = `https://registry.npmjs.org/${ encodeURIComponent(name) }/${versionToGet}`;
+        debug(`GET: ${url}`);
+
+        const response = await axios.get(url);
         const packageInfo = response.data;
+
+        debug(`Latest version number for package ${name}: ${packageInfo.version}`);
         return packageInfo.version;
     }
 
     // The main render function.
     function renderTemplate(templateName: string): void {
         const template = fs.readFileSync(path.join(templateDirectory, templateName)).toString();
+        debug(`${templateName}: Template contents read.`);
+
         const rendered = Mustache.render(template, variables);
+        debug(`${templateName}: Template rendered.`);
+
         const fileName = templateName.replace('.mustache', '');
-        fs.writeFileSync(path.join(variables.packageName, fileName), rendered);
+        const writePath = path.join(variables.packageName, fileName);
+        debug(`${templateName}: Writing rendered template to ${writePath}.`);
+        fs.writeFileSync(writePath, rendered);
     }
 
     // Render all the templates in the template folder.
+    debug(`Reading template directory: ${templateDirectory}`);
     const templates = fs.readdirSync(templateDirectory);
+    debug(`Templates directory read, starting render for ${templates.length} templates.`);
     for (const template of templates) {
         renderTemplate(template);
     }
 
     // Create standard folders.
+    debug(`Creating "dist" folder.`);
     fs.mkdirSync(path.join(packageName, 'src'));
-    fs.mkdirSync(path.join(packageName, 'dist'));
 
+    const indexTsPath = path.join(packageName, 'src', 'index.ts');
+    debug(`Writing empty "index.ts" file to ${indexTsPath}`);
     fs.writeFileSync(path.join(packageName, 'src', 'index.ts'), '');
+
+    debug(`Creating "dist" folder.`);
+    fs.mkdirSync(path.join(packageName, 'dist'));
 
 })();